register.h: Changed declaration of getting/setting register indices

+++ register.h
@@ -18,8 +18,8 @@ registerFile[0] = 0;
 void initPorts (void);
 void initRegFile (char *regFile);
 
-short int getRegIndex (InstructionType instr, int opNum);
-void setRegIndex (InstructionType instr, int opNum, char *regWord);
+short int getRegIndex (InstructionType instr, short int opNum);
+void setRegIndex (InstructionType instr, short int opNum, char *regWord);
 
 short int readRegValue (short int index);
 void writeRegValue (short int index, short int value);
